refactor(chrome-ext): seperate the Google Docs-specific content script logic
